refactor(gemini): rename and replace ProcessRateLimits trait with ProcessesRateLimits
- Renamed the `ProcessRateLimits` trait to `ProcessesRateLimits` for consistency..
