+++ src/Providers/Gemini/Gemini.php
@@ -18,7 +18,7 @@ use Prism\Prism\Exceptions\PrismProviderOverloadedException;
 use Prism\Prism\Exceptions\PrismRateLimitedException;
 use Prism\Prism\Images\Request as ImagesRequest;
 use Prism\Prism\Images\Response as ImagesResponse;
-use Prism\Prism\Providers\Gemini\Concerns\ProcessRateLimits;
+use Prism\Prism\Providers\Gemini\Concerns\ProcessesRateLimits;
 use Prism\Prism\Providers\Gemini\Handlers\Audio;
 use Prism\Prism\Providers\Gemini\Handlers\Cache;
 use Prism\Prism\Providers\Gemini\Handlers\Embeddings;
@@ -37,7 +37,7 @@ use Prism\Prism\ValueObjects\Messages\SystemMessage;
 class Gemini extends Provider
 {
     use InitializesClient;
-    use ProcessRateLimits;
+    use ProcessesRateLimits;
 
     public function __construct(
         #[\SensitiveParameter] public readonly string $apiKey,
